move listVmWithLocalData out of diskutils
listVmWithLocalData iterates over objects created in cloudinit package.
It is not a general purpose function for ephemeral-disks, so it does not
belong in ephemeraldiskutils.

Since listVmWithLocalData is used only by tests, I considered moving it
to cloud-init_test.go, but if feels wrong to expose low-level directory
structure to the tests.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>